# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "our_diary with Swagger",
    "description": "A simple CRUD API application made with Express and documented with Swagger",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3003"
    }
  ],
...
  "paths": {
...
    "/groups/": {
      "post": {
        "tags": [
          "GROUP"
        ],
        "summary": "GR_01 새로운 모임 생성",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "1": {
                    "value": {
                      "message": "성공",
                      "code": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "BadRequest": {
                    "value": {
                      "message": "입력이 유효하지 않습니다.",
                      "code": -1001
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "UnAuth": {
                    "value": {
                      "message": "미인증 회원, jwt 토큰이 없거나 만료됨",
                      "code": -1002
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "금지됨",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "WrongJwt": {
                    "value": {
                      "message": "잘못된 jwt 입니다",
                      "code": -5
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "내부 서버 오류",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "BackEnd": {
                    "value": {
                      "message": "서버 내부 오류",
                      "code": -1000,
                      "data": "에러 메세지"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "gr_emogi": {
                    "example": "any"
                  },
                  "fr_emogi": {
                    "example": "any"
                  },
                  "fr_color": {
                    "example": "any"
                  },
                  "gr_name": {
                    "example": "any"
                  },
                  "gr_search": {
                    "example": "any"
                  },
                  "fr_nickname": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "GROUP"
        ],
        "summary": "GR_02 모임 정보 수정",
        "description": "",
        "responses": {
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "BadRequest": {
                    "value": {
                      "message": "입력이 유효하지 않습니다.",
                      "code": -1001
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "UnAuth": {
                    "value": {
                      "message": "미인증 회원, jwt 토큰이 없거나 만료됨",
                      "code": -1002
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "금지됨",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "WrongJwt": {
                    "value": {
                      "message": "잘못된 jwt 입니다",
                      "code": -5
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "내부 서버 오류",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "BackEnd": {
                    "value": {
                      "message": "서버 내부 오류",
                      "code": -1000,
                      "data": "에러 메세지"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "emogi": {
                    "example": "any"
                  },
                  "search": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}